+++ src/cargo.rs
@@ -2,6 +2,8 @@ use crate::config::*;
 use crate::errors::RunError;
 use crate::path_utils::get_source_walker;
 use cargo_metadata::{diagnostic::DiagnosticLevel, CargoOpt, Message, Metadata, MetadataCommand};
+use lazy_static::lazy_static;
+use regex::Regex;
 use serde::{Deserialize, Serialize};
 use std::collections::{HashMap, HashSet};
 use std::env;
@@ -14,6 +16,29 @@ use std::process::{Command, Stdio};
 use toml::Value;
 use tracing::{error, trace, warn};
 use walkdir::{DirEntry, WalkDir};
+
+#[derive(Copy, Clone, Debug, Eq, PartialEq, Hash, Ord, PartialOrd)]
+enum Channel {
+    Stable,
+    Beta,
+    Nightly,
+}
+
+#[derive(Clone, Debug, Eq, PartialEq, Hash, Ord, PartialOrd)]
+struct CargoVersionInfo {
+    major: usize,
+    minor: usize,
+    channel: Channel,
+    year: usize,
+    month: usize,
+    day: usize,
+}
+
+impl CargoVersionInfo {
+    fn supports_llvm_cov(&self) -> bool {
+        self.minor >= 50 && self.channel == Channel::Nightly
+    }
+}
 
 #[derive(Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Deserialize, Serialize)]
 pub struct TestBinary {
@@ -97,6 +122,46 @@ impl DocTestBinaryMeta {
     }
 }
 
+lazy_static! {
+    static ref CARGO_VERSION_INFO: Option<CargoVersionInfo> = {
+        let version_info = Regex::new(
+            r"cargo (\d)\.(\d+)\.\d+([\-betanightly]*) \([[:alnum:]]+ (\d{4})-(\d{2})-(\d{2})\)",
+        )
+        .unwrap();
+        Command::new("cargo")
+            .arg("--version")
+            .output()
+            .map(|x| {
+                let s = String::from_utf8_lossy(&x.stdout);
+                if let Some(cap) = version_info.captures(&s) {
+                    let major = cap[1].parse().unwrap();
+                    let minor = cap[2].parse().unwrap();
+                    // We expect a string like `cargo 1.50.0-nightly (a0f433460 2020-02-01)
+                    // the version number either has `-nightly` `-beta` or empty for stable
+                    let channel = match &cap[3] {
+                        "-nightly" => Channel::Nightly,
+                        "-beta" => Channel::Beta,
+                        _ => Channel::Stable,
+                    };
+                    let year = cap[4].parse().unwrap();
+                    let month = cap[5].parse().unwrap();
+                    let day = cap[6].parse().unwrap();
+                    Some(CargoVersionInfo {
+                        major,
+                        minor,
+                        channel,
+                        year,
+                        month,
+                        day,
+                    })
+                } else {
+                    None
+                }
+            })
+            .unwrap_or(None)
+    };
+}
+
 pub fn get_tests(config: &Config) -> Result<Vec<TestBinary>, RunError> {
     let mut result = vec![];
     let manifest = match config.manifest.as_path().to_str() {
@@ -159,7 +224,8 @@ fn run_cargo(
                 }
                 Ok(Message::CompilerMessage(m)) => match m.message.level {
                     DiagnosticLevel::Error | DiagnosticLevel::Ice => {
-                        error = Some(RunError::TestCompile(m.message.message));
+                        let msg = format!("{}: {}", m.target.name, m.message.message);
+                        error = Some(RunError::TestCompile(msg));
                         break;
                     }
                     _ => {}
@@ -200,7 +266,6 @@ fn run_cargo(
             .filter_map(|e| e.ok())
             .filter(|e| matches!(e.metadata(), Ok(ref m) if m.is_file() && m.len() != 0))
             .collect::<Vec<_>>();
-
         let should_panics = get_panic_candidates(&dir_entries, config);
         for dt in &dir_entries {
             let mut tb = TestBinary::new(dt.path().to_path_buf(), ty);
@@ -244,7 +309,7 @@ fn convert_to_prefix(p: &Path) -> Option<String> {
 
 fn is_prefix_match(prefix: &str, entry: &Path) -> bool {
     convert_to_prefix(entry)
-        .map(|s| prefix.ends_with(&s))
+        .map(|s| s.contains(prefix))
         .unwrap_or(false)
 }
 
@@ -270,7 +335,6 @@ fn get_panic_candidates(tests: &[DirEntry], config: &Config) -> HashMap<String, 
                         if is_prefix_match(&test_binary.prefix, &p) && !checked_files.contains(path)
                         {
                             checked_files.insert(path.to_path_buf());
-                            trace!("Assessing {} for `should_panic` doctests", path.display());
                             let lines = find_panics_in_file(path).unwrap_or_default();
                             if !result.contains_key(&test_binary.prefix) {
                                 result.insert(test_binary.prefix.clone(), lines);
@@ -380,6 +444,10 @@ fn init_args(test_cmd: &mut Command, config: &Config) {
         test_cmd.arg("--profile");
         test_cmd.arg(profile);
     }
+    if let Some(jobs) = config.jobs {
+        test_cmd.arg("--jobs");
+        test_cmd.arg(jobs.to_string());
+    }
     if let Some(features) = config.features.as_ref() {
         test_cmd.arg("--features");
         test_cmd.arg(features);
@@ -450,6 +518,16 @@ fn clean_doctest_folder<P: AsRef<Path>>(doctest_dir: P) {
     }
 }
 
+fn handle_llvm_flags(value: &mut String, config: &Config) {
+    if (config.engine == TraceEngine::Auto || config.engine == TraceEngine::Llvm)
+        && supports_llvm_coverage()
+    {
+        value.push_str("-Z instrument-coverage ");
+    } else if config.engine == TraceEngine::Llvm {
+        error!("unable to utilise llvm coverage, due to compiler support. Falling back to Ptrace");
+    }
+}
+
 pub fn rustdoc_flags(config: &Config) -> String {
     const RUSTDOC: &str = "RUSTDOCFLAGS";
     let common_opts = " -C link-dead-code -C debuginfo=2 --cfg=tarpaulin ";
@@ -463,6 +541,7 @@ pub fn rustdoc_flags(config: &Config) -> String {
             value.push_str(vtemp.as_ref());
         }
     }
+    handle_llvm_flags(&mut value, config);
     value
 }
 
@@ -551,15 +630,20 @@ pub fn rust_flags(config: &Config) -> String {
     const RUSTFLAGS: &str = "RUSTFLAGS";
     let mut value = " -C link-dead-code -C debuginfo=2 ".to_string();
     if !config.avoid_cfg_tarpaulin {
-        value = format!("{}--cfg=tarpaulin ", value);
+        value.push_str("--cfg=tarpaulin ");
     }
     if config.release {
-        value = format!("{}-C debug-assertions=off ", value);
+        value.push_str("-C debug-assertions=off ");
     }
+    handle_llvm_flags(&mut value, config);
     if let Ok(vtemp) = env::var(RUSTFLAGS) {
         value.push_str(vtemp.as_ref());
     } else {
         value.push_str(gather_config_rust_flags(config).as_ref());
+        lazy_static! {
+            static ref DEBUG_INFO: Regex = Regex::new(r#"\-C\s*debuginfo=\d"#).unwrap();
+        }
+        value.push_str(&DEBUG_INFO.replace_all(&vtemp, " "));
     }
     value
 }
@@ -574,4 +658,50 @@ fn setup_environment(cmd: &mut Command, config: &Config) {
     let value = rustdoc_flags(config);
     trace!("Setting RUSTDOCFLAGS='{}'", value);
     cmd.env(rustdoc, value);
+}
+
+fn supports_llvm_coverage() -> bool {
+    if let Some(version) = CARGO_VERSION_INFO.as_ref() {
+        version.supports_llvm_cov()
+    } else {
+        false
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn llvm_cov_compatible_version() {
+        let version = CargoVersionInfo {
+            major: 1,
+            minor: 50,
+            channel: Channel::Nightly,
+            year: 2020,
+            month: 12,
+            day: 22,
+        };
+        assert!(version.supports_llvm_cov());
+    }
+
+    #[test]
+    fn llvm_cov_incompatible_version() {
+        let mut version = CargoVersionInfo {
+            major: 1,
+            minor: 48,
+            channel: Channel::Stable,
+            year: 2020,
+            month: 10,
+            day: 14,
+        };
+        assert!(!version.supports_llvm_cov());
+        version.channel = Channel::Beta;
+        assert!(!version.supports_llvm_cov());
+        version.minor = 50;
+        assert!(!version.supports_llvm_cov());
+        version.minor = 58;
+        version.channel = Channel::Stable;
+        assert!(!version.supports_llvm_cov());
+    }
 }