+++ src/config/mod.rs
@@ -163,6 +163,10 @@ pub struct Config {
     pub color: Color,
     /// Follow traced executables down
     pub follow_exec: bool,
+    /// Number of jobs used for building the tests
+    pub jobs: Option<usize>,
+    /// Engine to use to collect coverage
+    pub engine: TraceEngine,
 }
 
 fn default_test_timeout() -> Duration {
@@ -223,7 +227,9 @@ impl Default for Config {
             fail_under: None,
             metadata: RefCell::new(None),
             avoid_cfg_tarpaulin: false,
+            jobs: None,
             color: Color::Auto,
+            engine: TraceEngine::Ptrace,
         }
     }
 }
@@ -293,9 +299,11 @@ impl<'a> From<&'a ArgMatches<'a>> for ConfigWrapper {
             bench_names: get_list(args, "bench").iter().cloned().collect(),
             example_names: get_list(args, "example").iter().cloned().collect(),
             fail_under: value_t!(args.value_of("fail-under"), f64).ok(),
+            jobs: value_t!(args.value_of("jobs"), usize).ok(),
             profile: get_profile(args),
             metadata: RefCell::new(None),
             avoid_cfg_tarpaulin: args.is_present("avoid-cfg-tarpaulin"),
+            engine: TraceEngine::Ptrace,
         };
         if args.is_present("ignore-config") {
             Self(vec![args_config])
@@ -505,6 +513,9 @@ impl Config {
         self.ignore_tests |= other.ignore_tests;
         self.no_fail_fast |= other.no_fail_fast;
 
+        if self.jobs.is_none() {
+            self.jobs = other.jobs;
+        }
         if self.fail_under.is_none()
             || other.fail_under.is_some() && other.fail_under.unwrap() < self.fail_under.unwrap()
         {
@@ -809,31 +820,39 @@ mod tests {
 
     #[test]
     fn relative_path_test() {
-        let path_a = Path::new("/this/should/form/a/rel/path/");
-        let path_b = Path::new("/this/should/form/b/rel/path/");
+        cfg_if::cfg_if! {
+            if #[cfg(windows)] {
+                let root_base = "C:";
+            } else {
+                let root_base = "";
+            }
+        }
+        let path_a = PathBuf::from(format!("{}/this/should/form/a/rel/path/", root_base));
+        let path_b = PathBuf::from(format!("{}/this/should/form/b/rel/path/", root_base));
 
-        let rel_path = path_relative_from(path_b, path_a);
+        let rel_path = path_relative_from(&path_b, &path_a);
         assert!(rel_path.is_some());
         assert_eq!(
-            rel_path.unwrap().to_str().unwrap(),
-            "../../../b/rel/path",
+            rel_path.unwrap(),
+            Path::new("../../../b/rel/path"),
             "Wrong relative path"
         );
 
-        let path_a = Path::new("/this/should/not/form/a/rel/path/");
-        let path_b = Path::new("./this/should/not/form/a/rel/path/");
-
-        let rel_path = path_relative_from(path_b, path_a);
+        let path_a = PathBuf::from(format!("{}/this/should/not/form/a/rel/path/", root_base));
+        let path_b = Path::new("this/should/not/form/a/rel/path/");
+        assert!(!path_b.is_absolute());
+        assert!(path_a.is_absolute());
+        let rel_path = path_relative_from(path_b, &path_a);
         assert_eq!(rel_path, None, "Did not expect relative path");
 
-        let path_a = Path::new("./this/should/form/a/rel/path/");
-        let path_b = Path::new("./this/should/form/b/rel/path/");
+        let path_a = Path::new("this/should/form/a/rel/path/");
+        let path_b = Path::new("this/should/form/b/rel/path/");
 
         let rel_path = path_relative_from(path_b, path_a);
         assert!(rel_path.is_some());
         assert_eq!(
-            rel_path.unwrap().to_str().unwrap(),
-            "../../../b/rel/path",
+            rel_path.unwrap(),
+            Path::new("../../../b/rel/path"),
             "Wrong relative path"
         );
     }